# swagger.json
{
  "swagger": "2.0.1",
  "info": {
    "title": "Fambase",
    "description": "The main objective of this project is to integrate the back end and front end of the application with minor changes in the front end and development of the admin panel.",
    "version": "0.0.1"
  },
  "produces": ["multipart/form-data", "application/json"],
  "host": "192.168.2.231:3013",
  "basePath": "",
  "paths": {
    "/auth/signup": {
      "post": {
        "operationId": "register",
        "tags": ["Auth"],
        "description": "signup user",
        "parameters": [
          {
            "name": "firstName",
            "in": "formData",
            "description": "Please Enter First Name.",
            "type": "string"
          },
          {
            "name": "lastName",
            "in": "formData",
            "description": "Please Enter Last Name.",
            "type": "string"
          },
          {
            "name": "email",
            "in": "formData",
            "description": "Please Enter Email.",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "description": "Please Enter password.",
            "required": true,
            "type": "string"
          },
          {
            "name": "birthday",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "gender",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "phoneNo",
            "in": "formData",
            "required": true,
            "type": "number"
          },
          {
            "name": "countryCode",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "userName",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/auth/verifyOtp": {
      "post": {
        "operationId": "verifyOtp",
        "tags": ["Auth"],
        "description": "verifyOtp user",
        "parameters": [
          {
            "name": "email",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "otp",
            "in": "formData",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {}
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "login",
        "tags": ["Auth"],
        "description": "login user",
        "parameters": [
          {
            "name": "email",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "userName",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/auth/getProfile": {
      "get": {
        "operationId": "getProfile",
        "tags": ["Auth"],
        "description": "getProfile",
        "parameters": [
          {
            "name": "x-token",
            "in": "header",
            "required": true,
            "type": "string",
            "value": "Bearer "
          }
        ],
        "responses": {}
      }
    },
    "/auth/updateProfile": {
      "put": {
        "operationId": "updateProfile",
        "tags": ["Auth"],
        "description": "updateProfile user",
        "parameters": [
          {
            "name": "x-token",
            "in": "header",
            "required": true,
            "type": "string",
            "value": "Bearer "
          },
          {
            "name":"id",
            "in":"formData",
            "type":"string"
          },
          {
            "name": "firstName",
            "in": "formData",
            "type": "string"
          },
          {
            "name": "lastName",
            "in": "formData",
          
            "type": "string"
          },
          {
            "name": "email",
            "in": "formData",
            "type": "string"
          },
         
          {
            "name": "birthday",
            "in": "formData",
            "type": "string"
          },
          {
            "name": "gender",
            "in": "formData",
            "type": "string"
          },
          {
            "name": "phoneNo",
            "in": "formData",
            "type": "string"
          },
          {
            "name": "countryCode",
            "in": "formData",
            "type": "string"
          },
          {
            "name": "userName",
            "in": "formData",
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/admin/userList": {
      "get": {
        "operationId": "userList",
        "tags": ["Admin"],
        "description": "userList",
        "parameters": [
          {
            "name": "x-token",
            "in": "header",
            "required": true,
            "type": "string",
            "value": "Bearer "
          },
          {
            "name":"pageNo",
            "in":"query",
            "type":"string"
          },
          {
            "name":"pageSize",
            "in":"query",
            "type":"string"
          },
          {
            "name":"search",
            "in":"query",
            "type":"string"
          }
        ],
        "responses": {}
      }
    },
    "/admin/userList/{id}": {
      "delete": {
        "operationId": "deleteUser",
        "tags": ["Admin"],
        "description": "deleteUser",
        "parameters": [
          {
            "name": "x-token",
            "in": "header",
            "required": true,
            "type": "string",
            "value": "Bearer "
          },
          {
            "name": "id",
            "in": "path",
            "description": "Please Enter userId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/admin/dashboard":{
      "get":{
        "operationId":"dashboard",
        "tags":["Admin"],
        "description":"Count Total Block and Active Users",
        "parameters":[
          {
            "name":"x-token",
            "in":"header",
            "required":true,
            "type":"string",
            "value":"Bearer "
          }
        ],
        "responses":{}
      }
    },
    "/auth/changePassword": {
      "put": {
        "operationId": "changePassword",
        "tags": ["Auth"],
        "description": "changePassword",
        "parameters": [
          {
            "name": "x-token",
            "in": "header",
            "required": true,
            "type": "string",
            "value": "Bearer "
          },
          {
            "name": "oldPassword",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/auth/forgotPassword": {
      "post": {
        "operationId": "forgotPassword",
        "tags": ["Auth"],
        "description": "forgotPassword",
        "parameters": [
          {
            "name": "email",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    }
  }
}